requirements.txt

pygame>=2.0.0
Flask>=2.0.0
Pillow>=8.0.0

